refactor: Split code from main
