+++ block.go
@@ -1,0 +1,34 @@
+package main
+
+import "time"
+
+type Block struct {
+	Timestamp     int64  // current timestamp
+	Data          []byte // actual valuable info
+	PrevBlockHash []byte // hash of previous block
+	Hash          []byte // hash of block
+	Nonce         int
+}
+
+// func (b *Block) SetHash() {
+// 	timestamp := []byte(strconv.FormatInt(b.Timestamp, 10))
+// 	headers := bytes.Join([][]byte{b.PrevBlockHash, b.Data, timestamp}, []byte{})
+// 	hash := sha256.Sum256(headers)
+
+// 	b.Hash = hash[:]
+// }
+
+func NewBlock(data string, prevBlockHash []byte) *Block {
+	block := &Block{time.Now().Unix(), []byte(data), prevBlockHash, []byte{}, 0}
+	pow := NewProofOfWork(block)
+	nonce, hash := pow.Run()
+
+	block.Hash = hash[:]
+	block.Nonce = nonce
+	return block
+}
+
+// The first block in the chain
+func NewGenesisBlock() *Block {
+	return NewBlock("Genesis Block", []byte{})
+}
