+++ main.go
@@ -3,53 +3,7 @@ package main
 import (
 	"fmt"
 	"strconv"
-	"time"
 )
-
-type Block struct {
-	Timestamp     int64  // current timestamp
-	Data          []byte // actual valuable info
-	PrevBlockHash []byte // hash of previous block
-	Hash          []byte // hash of block
-	Nonce         int
-}
-
-// func (b *Block) SetHash() {
-// 	timestamp := []byte(strconv.FormatInt(b.Timestamp, 10))
-// 	headers := bytes.Join([][]byte{b.PrevBlockHash, b.Data, timestamp}, []byte{})
-// 	hash := sha256.Sum256(headers)
-
-// 	b.Hash = hash[:]
-// }
-
-func NewBlock(data string, prevBlockHash []byte) *Block {
-	block := &Block{time.Now().Unix(), []byte(data), prevBlockHash, []byte{}, 0}
-	pow := NewProofOfWork(block)
-	nonce, hash := pow.Run()
-
-	block.Hash = hash[:]
-	block.Nonce = nonce
-	return block
-}
-
-type Blockchain struct {
-	blocks []*Block
-}
-
-func (bc *Blockchain) AddBlock(data string) {
-	prevBlock := bc.blocks[len(bc.blocks)-1]
-	newBlock := NewBlock(data, prevBlock.Hash)
-	bc.blocks = append(bc.blocks, newBlock)
-}
-
-// The first block in the chain
-func NewGenesisBlock() *Block {
-	return NewBlock("Genesis Block", []byte{})
-}
-
-func NewBlockchain() *Blockchain {
-	return &Blockchain{[]*Block{NewGenesisBlock()}}
-}
 
 func main() {
 	bc := NewBlockchain()
